A little bit of touch-up work
Removed some unused temporary models, added the beam tower icon to the UI, raised the beam tower model so it wouldn't spawn in the ground, made the building UI a little bit smaller so it fits the buttons better. Made it so that when you build a tower, the tower info menu wouldn't show up because it was bugged when showing up immediately after building.